Implemented more creation validation (value-ranges and custom).

+++ Solution/Fabric.New.Api.Objects/Meta/FabSpecValue.cs
@@ -8,8 +8,8 @@
 		public string Description { get; set; }
 
 		public bool? IsOptional { get; set; }
-		public int? Min { get; set; }
-		public int? Max { get; set; }
+		public long? Min { get; set; }
+		public long? Max { get; set; }
 		public int? LenMin { get; set; }
 		public int? LenMax { get; set; }
 		public string ValidRegex { get; set; }
